Implement MatchCache for RelativeFragment matches
Whenever a RelativeFragment is matched it will register its unique ID in
the cache, so that its siblings can reference the cache to see if they
should render or not. This allows us to implement a greedy matching
algorithm that renders all RelativeFragment instances that match the
first route matched when their parent was renderered.