Integrate MatchCache with RelativeFragment
Whenever a fragment renders it checks to see if a parentId exists and
checks the cache status before rendering. If the cache states that the
route has already been matched for that parentId, and the matched route
is not equal to the route of the fragment, then the fragment returns
null.