Parsing the query string on location change
There was an issue when using the back button to navigate to a route
that had a query string in it. It would be available in the search
field, but it was not available as a query object.

This commit simply performs the same parsing of the query string that
the ROUTER_POP state receives.